wip: #16 Migrated to aircrack-ng
